created common interface for debugService and webSocketDebugService
